advanced-api-project/requirements.txt
django==4.2
djangorestframework==3.16.1
mysql-connector-python==9.4.0
python-dotenv==1.0.1